Move Admin layout to layouts folder
